@@ -1,14 +1,15 @@
 {
   "name": "react-native-animatable",
-  "version": "1.2.3",
+  "version": "1.3.3",
   "description": "Easy to use declarative transitions and animations for React Native",
+  "typings": "typings/react-native-animatable.d.ts",
   "main": "index.js",
   "scripts": {
     "jest": "./node_modules/.bin/jest",
     "jest:watch": "npm run jest -- --watch",
     "lint": "./node_modules/.bin/eslint ./*.js",
     "test": "npm run lint && npm run jest",
-    "format": "./node_modules/.bin/prettier --single-quote --trailing-comma all --write {,definitions/,__tests__}*.js"
+    "format": "./node_modules/.bin/prettier --write {,definitions/,__tests__/}*.js"
   },
   "keywords": [
     "react-native",
@@ -40,30 +41,37 @@
   },
   "license": "MIT",
   "jest": {
-    "preset": "jest-react-native",
+    "preset": "react-native",
     "modulePathIgnorePatterns": [
-      "<rootDir>/Example/"
-    ]
+      "<rootDir>/Examples/"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/Examples/"
+    ],
+    "collectCoverage": true,
+    "coverageDirectory": "<rootDir>/coverage/"
   },
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-eslint": "^7.0.0",
-    "babel-jest": "^20.0.3",
-    "babel-preset-react-native": "^1.9.0",
-    "eslint": "^3.7.1",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-react": "^7.0.1",
-    "jest": "20.0.4",
-    "jest-cli": "20.0.4",
-    "jest-react-native": "18.0.0",
-    "prettier": "^1.3.1",
-    "react": "*",
-    "react-native": "*",
-    "react-test-renderer": "15.5.4"
+    "@babel/core": "^7.6.4",
+    "@babel/runtime": "^7.6.3",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-preset-react-native": "4.0.0",
+    "eslint": "6.1.0",
+    "eslint-config-airbnb": "18.0.1",
+    "eslint-config-prettier": "^6.4.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "jest": "^24.9.0",
+    "prettier": "^1.18.2",
+    "react": "16.9.0",
+    "react-native": "0.61.2",
+    "react-test-renderer": "16.9.0"
   },
   "dependencies": {
-    "prop-types": "^15.5.10"
+    "prop-types": "^15.7.2"
   }
 }
